refactor: admin과 client에서 사용하는 공통 상수 shared 에서 관리
