+++ apps/admin/app/(auth)/(attendance)/attendance/[memberId]/[sessionId]/_components/attendance-modify-status.tsx
@@ -17,14 +17,7 @@ import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { useState } from 'react';
 import { Profile } from '@/components/attendance/profile';
 import { modifyAttendanceStatusOptions } from '@/remotes/mutations/attendance';
-
-const ATTENDANCE_STATUS = [
-	{ label: '출석', value: 'PRESENT' },
-	{ label: '결석', value: 'ABSENT' },
-	{ label: '지각', value: 'LATE' },
-	{ label: '인정', value: 'EXCUSED_ABSENT' },
-	{ label: '조퇴', value: 'EARLY_LEAVE' },
-];
+import { ATTENDANCE_STATUS_OPTIONS } from '@dpm-core/shared';
 
 interface AttendanceModifyStatusProps {
 	sessionId: number;
@@ -75,7 +68,7 @@ export const AttendanceModifyStatus = (props: AttendanceModifyStatusProps) => {
 						className="w-full mt-3 flex border border-line-normal rounded-lg"
 						onValueChange={(value) => setSelectedStatus(value as AttendanceStatus)}
 					>
-						{ATTENDANCE_STATUS.map((status) => (
+						{ATTENDANCE_STATUS_OPTIONS.map((status) => (
 							<RadioGroup.Item
 								key={status.value}
 								value={status.value}
