--- packages/shared/src/constants/attendance.ts
@@ -1,3 +1,0 @@
-export const ONE_MINUTE = 1000 * 60;
-export const ATTENDANCE_GAP_DURATION = ONE_MINUTE * 15; // 15 minutes
-export const ATTENDANCE_LATE_DURATION = ONE_MINUTE * 15; // 15 minutes
